Move test utilities into its own crate
