+++ engine/snapshot/Cargo.toml
@@ -18,3 +18,6 @@ features = [ "random", "chacha", "hmac", "sha" ]
 [dev-dependencies]
 tempfile = "3.1.0"
 rand = "0.7.3"
+
+  [dev-dependencies.stronghold-test-utils]
+  path = "../../test_utils"
