+++ engine/snapshot/src/lib.rs
@@ -24,9 +24,6 @@
 pub mod files;
 pub mod kdf;
 
-#[cfg(test)]
-mod test_utils;
-
 mod logic;
 pub use logic::*;
