+++ engine/vault/Cargo.toml
@@ -15,13 +15,14 @@ json = "0.12"
 rand = "0.7.3"
 criterion = "0.3.3"
 
-[dev-dependencies.random]
-path = "../random"
-version = "0.2"
+  [dev-dependencies.random]
+  path = "../random"
 
-[dev-dependencies.crypto]
-path = "../crypto"
-version = "0.2"
+  [dev-dependencies.crypto]
+  path = "../crypto"
+
+  [dev-dependencies.stronghold-test-utils]
+  path = "../../test_utils"
 
 [[bench]]
 name = "benchmark"
