+++ test_utils/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "stronghold-test-utils"
+version = "0.1.0"
+authors = [ "Gustav Behm <gustav.behm@iota.org>" ]
+edition = "2018"
+license = "Apache-2.0"
+
+[lib]
+name = "test_utils"
+
+[dependencies]
+rand = "0.7.3"
