+++ test_utils/src/lib.rs
@@ -1,4 +1,4 @@
-// Copyright 2020 IOTA Stiftung
+// Copyright 2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
 use std::{
@@ -7,31 +7,7 @@ use std::{
     path::Path,
 };
 
-pub mod fresh {
-    pub fn bytestring() -> Vec<u8> {
-        let s = if rand::random::<u8>() % 4 == 0 {
-            0
-        } else {
-            rand::random::<usize>() % 4096
-        };
-
-        let mut bs = Vec::with_capacity(s);
-        for _ in 1..s {
-            bs.push(rand::random());
-        }
-        bs
-    }
-
-    #[allow(dead_code)]
-    pub fn non_empty_bytestring() -> Vec<u8> {
-        let s = (rand::random::<usize>() % 4096) + 1;
-        let mut bs = Vec::with_capacity(s);
-        for _ in 1..s {
-            bs.push(rand::random());
-        }
-        bs
-    }
-}
+pub mod fresh;
 
 pub fn corrupt(bs: &mut [u8]) {
     if bs.is_empty() {
@@ -47,7 +23,7 @@ pub fn corrupt(bs: &mut [u8]) {
     }
 }
 
-pub fn seek_to_beginning(f: &mut File) {
+fn seek_to_beginning(f: &mut File) {
     f.seek(SeekFrom::Start(0)).unwrap();
 }
